Moving app file
