+++ components/openai/actions/chat/chat.mjs
@@ -1,10 +1,10 @@
-import openai from "../../app/openai.app.mjs";
+import openai from "../../openai.app.mjs";
 import common from "../common/common.mjs";
 
 export default {
   ...common,
   name: "Chat",
-  version: "0.1.2",
+  version: "0.1.3",
   key: "openai-chat",
   description: "The Chat API, using the `gpt-3.5-turbo` or `gpt-4` model. [See docs here](https://platform.openai.com/docs/api-reference/chat)",
   type: "action",
