+++ components/openai/actions/classify-items-into-categories/classify-items-into-categories.mjs
@@ -3,7 +3,7 @@ import common from "../common/common-helper.mjs";
 export default {
   ...common,
   name: "Classify Items into Categories",
-  version: "0.0.4",
+  version: "0.0.5",
   key: "openai-classify-items-into-categories",
   description: "Classify items into specific categories using the Chat API",
   type: "action",
