+++ components/openai/actions/create-embeddings/create-embeddings.mjs
@@ -1,10 +1,10 @@
-import openai from "../../app/openai.app.mjs";
+import openai from "../../openai.app.mjs";
 import common from "../common/common.mjs";
 import { ConfigurationError } from "@pipedream/platform";
 
 export default {
   name: "Create Embeddings",
-  version: "0.0.2",
+  version: "0.0.3",
   key: "openai-create-embeddings",
   description: "Get a vector representation of a given input that can be easily consumed by machine learning models and algorithms. [See the docs here](https://platform.openai.com/docs/api-reference/embeddings)",
   type: "action",
