+++ components/openai/actions/send-prompt/send-prompt.mjs
@@ -1,10 +1,10 @@
-import openai from "../../app/openai.app.mjs";
+import openai from "../../openai.app.mjs";
 import common from "../common/common.mjs";
 
 export default {
   ...common,
   name: "Create Completion (Send Prompt)",
-  version: "0.1.1",
+  version: "0.1.2",
   key: "openai-send-prompt",
   description: "OpenAI recommends using the **Chat** action for the latest `gpt-3.5-turbo` API, since it's faster and 10x cheaper. This action creates a completion for the provided prompt and parameters using the older `/completions` API. [See docs here](https://beta.openai.com/docs/api-reference/completions/create)",
   type: "action",
