+++ components/openai/actions/summarize/summarize.mjs
@@ -3,7 +3,7 @@ import common from "../common/common-helper.mjs";
 export default {
   ...common,
   name: "Summarize Text",
-  version: "0.0.4",
+  version: "0.0.5",
   key: "openai-summarize",
   description: "Summarizes text using the Chat API",
   type: "action",
